third_party/blink/web_tests/platform/generic/fast/inline/inline-relative-offset-boundingbox-expected.txt

Bug 91168: REGRESSION: LayoutInline boundingBox ignores relative position offset

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS inlineRect.left is parentRect.left + inlineLeftOffset
